refactor (merge_sort): simplify merge
Create two subarray and fill the original array with the sorted
values. This solution feels simpler than copying the original
array into a new one, and then fill it with the sorted values,
and copy them over to the original array. The simplification has
to do with the loop end conditions which become simpler when they
are stated in terms of the length of the subarray, rather than a
last value in the original array.